Fixes memory leaks in sprite processing
Removes leftover FIXME comments related to the issue.

My theory as to why this was happening is the compiler/debugger didn't know how to interpret the free 'signal'. why this happens is beyond me. what I this to solve this was, I changed the GDB that was being used to the strawberry one.  If this pops in again I change the gdb to minGW/bin/gdb.exe or something